[#57] baseIteratee의 type 파일 분리
